Add Ustar header fields.
Adds support for the additional fields of Ustar format. A type
Header.compatibility indicates which version of tar should be assumed by
functions requiring it, in particular Header.Link.of_int defaults to
previous V7 behaviour. ?level added to functions which need to be
determine compatibility level to use. Extra fields are not read in this
commit.