Support for writing Ustar headers.
Ustar format support allowing non-zero devmajor and devminor and
non-empty uname and gname in the header. file_name can also be up to 256
characters long (255 characters for a directory name), providing that it
can be split at a path-separator such that the basename is 100
characters or fewer and the dirname 155 characters or fewer.

?level added to Archive.create_gen to allow writing Ustar headers.