Support unmarshalling pax headers
Pax headers are created when some ustar header fields have values
too big to represent, such as large user or group ids. Critically
the spec says that if a pax field is present, the regular field
should be ignored and unfortunately some tools exploit this by
writing garbage to the regular header.

This patch extends the unmarshalling code to read pax headers.

Signed-off-by: David Scott <dave@recoil.org>