Don't print the name of the file in extract functions
IMO this is an undocumented and ugly side-effect of the function. The
library shouldn't write anything to standard output! Or at least pass
the channel through an optional argument.
As the side-effect can be achieved by the `dest` function provided by
the user of the library, we can remove that printing code.

Signed-off-by: Antonin Décimo <antonin@tarides.com>