Remove Header module from the functor and use only Tar.Header
The Header module doesn't depend on anything in the Tar.Make functor,
as it has nothing to do with IO but simply to map the tar header to an
OCaml record type and associated helper functions.

Removing it from the functor helps cleaning the code an sharing the
header type definition between different IO backends.

Signed-off-by: Antonin Décimo <antonin@tarides.com>