Tar_mirage: Implement remove
This implementation of remove only works if the key is the very last.
Otherwise ``Error `Append_only`` is returned.

The reason is because in that case we simply overwrite the header and
the next block with zeroes. In the other case we would have to perform
compaction of the trailing files which could be very expensive and is
very error prone.